Add function which is update payment status (PENDING - PAID - CANCELLED)

+++ KoiVeterinaryServiceCenter_Server/KoiVeterinaryServiceCenter.DataAccess/IRepository/IOrderItemsRepository.cs
@@ -5,6 +5,6 @@ namespace KoiVeterinaryServiceCenter.DataAccess.IRepository
     public interface IOrderItemsRepository : IRepository<OrderItems>
     {
         void Update(OrderItems orderItems);
-        Task<OrderItems> GetById(int id);
+        Task<OrderItems> GetById(long id);
     }
 }
